refactor: removed AddEventModel and replaced it with form component
